实现了JDBC基本功能

+++ src/com/sec/entity/IdEntity.java
@@ -4,13 +4,13 @@ package com.sec.entity;
  * Created by Administrator on 2016/11/26.
  */
 public abstract class IdEntity {
-    private Long id;
+    private int id;
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public int getId() {
 
         return id;
     }
